refactor: Change sidebar to general sidebar
