+++ src/components/Sidebar/NavigationSwitcher.tsx
